収入部分のルーティング修正
