+++ app/controllers/concerns/check_user_login_status.rb
@@ -1,0 +1,9 @@
+module CheckUserLoginStatus
+  def logged_in_user
+    unless logged_in?
+      store_location
+      flash[:danger] = "ログインしてください"
+      redirect_to login_url, status: :see_other
+    end
+  end
+end
